ENHANCE: tree manipulation functions
